requirements.txt
groq==0.4.0
duckduckgo-search==3.9.3
python-dotenv==1.0.0
google-api-python-client